+++ tests/integration/components/sweet-alert-test.js
@@ -1,6 +1,6 @@
 import { module, test } from 'qunit';
 import { setupRenderingTest } from 'ember-qunit';
-import { render } from 'ember-test-helpers';
+import { render } from '@ember/test-helpers';
 import hbs from 'htmlbars-inline-precompile';
 import { open, confirmAndClose, cancelAndClose } from 'ember-sweetalert/test-support';